+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Latest
 
+- Added all values to the Species.Category choices.
+
 ## 0.2.1
 
 - Added timestamp fields to track when records are added.